Don't try to kill if pid is nil

+++ lib/spring/application_manager.rb
@@ -59,7 +59,7 @@ module Spring
     end
 
     def stop
-      Process.kill('TERM', pid)
+      Process.kill('TERM', pid) if pid
     end
 
     private
